[move/#113] 전체 일정 파일 이동
